+++ ios/ScreenShareAccPackKit/OTScreenShareAccPackKit/ReceiveScreenViewController.m
@@ -7,9 +7,10 @@
 //
 
 #import "ReceiveScreenViewController.h"
-#import <OTScreenShareKit/OTScreenShareKit.h>
+#import "AppDelegate.h"
+#import "OTScreenSharer.h"
 
-@interface ReceiveScreenViewController ()
+@interface ReceiveScreenViewController () <OTScreenShareDataSource>
 @property (nonatomic) OTScreenSharer *screenSharer;
 @end
 
@@ -21,17 +22,18 @@
     UIBarButtonItem *previewBarButtonItem = [[UIBarButtonItem alloc] initWithTitle:@"Navigate" style:UIBarButtonItemStylePlain target:self action:@selector(navigateToOtherViews)];
     self.navigationItem.rightBarButtonItem = previewBarButtonItem;
     
-    self.screenSharer = [OTScreenSharer sharedInstance];
+    self.screenSharer = [[OTScreenSharer alloc] init];
+    self.screenSharer.dataSource = self;
     [self.screenSharer connectWithView:nil
-                         handler:^(OTScreenShareSignal signal, NSError *error) {
+                         handler:^(OTCommunicationSignal signal, NSError *error) {
                              
                              if (!error) {
                                  
-                                 if (signal == OTScreenShareSignalSessionDidConnect) {
+                                 if (signal == OTPublisherCreated) {
                                      self.screenSharer.publishAudio = NO;
                                      self.screenSharer.subscribeToAudio = NO;
                                  }
-                                 else if (signal == OTScreenShareSignalSubscriberDidConnect) {
+                                 else if (signal == OTSubscriberReady) {
                                      
                                      [self.screenSharer.subscriberView removeFromSuperview];
                                      self.screenSharer.subscriberView.frame = self.view.bounds;
@@ -39,6 +41,11 @@
                                  }
                              }
                          }];
+}
+
+- (void)viewWillDisappear:(BOOL)animated {
+    [super viewWillDisappear:animated];
+    [self.screenSharer disconnect];
 }
 
 - (void)navigateToOtherViews {
@@ -50,11 +57,11 @@
         
         if (self.screenSharer.isScreenSharing) {
             
-            if (self.screenSharer.subscriberVideoContentMode == OTScreenShareVideoViewFit) {
-                self.screenSharer.subscriberVideoContentMode = OTScreenShareVideoViewFill;
+            if (self.screenSharer.subscriberVideoContentMode == OTVideoViewFit) {
+                self.screenSharer.subscriberVideoContentMode = OTVideoViewFill;
             }
             else {
-                self.screenSharer.subscriberVideoContentMode = OTScreenShareVideoViewFit;
+                self.screenSharer.subscriberVideoContentMode = OTVideoViewFit;
             }
         }
     }]];
@@ -65,6 +72,10 @@
     }]];
     
     [self presentViewController:alert animated:YES completion:nil];
+}
+
+- (OTAcceleratorSession *)sessionOfOTScreenSharer:(OTScreenSharer *)screenSharer {
+    return [(AppDelegate*)[[UIApplication sharedApplication] delegate] getSharedAcceleratorSession];
 }
 
 @end