+++ ios/ScreenShareAccPackKit/OTScreenShareAccPackKit/ShareWholeScreenViewController.m
@@ -8,9 +8,10 @@
 
 #import "ShareWholeScreenViewController.h"
 #import "ColorViewController.h"
-#import <OTScreenShareKit/OTScreenShareKit.h>
+#import "AppDelegate.h"
+#import "OTScreenSharer.h"
 
-@interface ShareWholeScreenViewController ()
+@interface ShareWholeScreenViewController () <OTScreenShareDataSource>
 @property (weak, nonatomic) IBOutlet UIWebView *webView;
 @property (nonatomic) UIColor *viewControllerColor;
 @property (nonatomic) OTScreenSharer *screenSharer;
@@ -28,12 +29,25 @@
     UIBarButtonItem *previewBarButtonItem = [[UIBarButtonItem alloc] initWithTitle:@"Navigate" style:UIBarButtonItemStylePlain target:self action:@selector(navigateToOtherViews)];
     self.navigationItem.rightBarButtonItem = previewBarButtonItem;
 
-    self.screenSharer = [OTScreenSharer sharedInstance];
+    self.screenSharer = [[OTScreenSharer alloc] init];
+    self.screenSharer.dataSource = self;
     [self.screenSharer connectWithView:[UIApplication sharedApplication].keyWindow.rootViewController.view
-                               handler:^(OTScreenShareSignal signal, NSError *error) {
+                               handler:^(OTCommunicationSignal signal, NSError *error) {
         
-                                   NSLog(@"%@", error);
+                                   if (signal == OTPublisherCreated) {
+                                       self.screenSharer.publisherView.frame = CGRectMake(10, 200, 100, 100);
+                                       [self.view addSubview:self.screenSharer.publisherView];
+                                   }
+                                   else if (signal == OTSubscriberReady) {
+                                       self.screenSharer.subscriberView.frame = CGRectMake(10, 80, 100, 100);
+                                       [self.view addSubview:self.screenSharer.subscriberView];
+                                   }
                                }];
+}
+
+- (void)viewWillDisappear:(BOOL)animated {
+    [super viewWillDisappear:animated];
+    [self.screenSharer disconnect];
 }
 
 - (void)navigateToOtherViews {
@@ -58,17 +72,31 @@
         [self performSegueWithIdentifier:@"ColorViewControllerSegue" sender:nil];
     }]];
     
-    [alert addAction:[UIAlertAction actionWithTitle:@"TURN ON/OFF AUDIO" style:UIAlertActionStyleDefault handler:^(UIAlertAction *action) {
+    [alert addAction:[UIAlertAction actionWithTitle:@"TURN ON/OFF PUBLISHER AUDIO" style:UIAlertActionStyleDefault handler:^(UIAlertAction *action) {
         
         if (self.screenSharer.isScreenSharing) {
             self.screenSharer.publishAudio = !self.screenSharer.publishAudio;
         }
     }]];
     
-    [alert addAction:[UIAlertAction actionWithTitle:@"TURN ON/OFF VIDEO" style:UIAlertActionStyleDefault handler:^(UIAlertAction *action) {
+    [alert addAction:[UIAlertAction actionWithTitle:@"TURN ON/OFF PUBLISHER VIDEO" style:UIAlertActionStyleDefault handler:^(UIAlertAction *action) {
         
         if (self.screenSharer.isScreenSharing) {
             self.screenSharer.publishVideo = !self.screenSharer.publishVideo;
+        }
+    }]];
+    
+    [alert addAction:[UIAlertAction actionWithTitle:@"TURN ON/OFF SUBSCRIBER AUDIO" style:UIAlertActionStyleDefault handler:^(UIAlertAction *action) {
+        
+        if (self.screenSharer.isScreenSharing) {
+            self.screenSharer.subscribeToAudio = !self.screenSharer.subscribeToAudio;
+        }
+    }]];
+    
+    [alert addAction:[UIAlertAction actionWithTitle:@"TURN ON/OFF SUBSCRIBER VIDEO" style:UIAlertActionStyleDefault handler:^(UIAlertAction *action) {
+        
+        if (self.screenSharer.isScreenSharing) {
+            self.screenSharer.subscribeToVideo = !self.screenSharer.subscribeToVideo;
         }
     }]];
     
@@ -76,11 +104,11 @@
         
         if (self.screenSharer.isScreenSharing) {
             
-            if (self.screenSharer.subscriberVideoContentMode == OTScreenShareVideoViewFit) {
-                self.screenSharer.subscriberVideoContentMode = OTScreenShareVideoViewFill;
+            if (self.screenSharer.subscriberVideoContentMode == OTVideoViewFit) {
+                self.screenSharer.subscriberVideoContentMode = OTVideoViewFill;
             }
             else {
-                self.screenSharer.subscriberVideoContentMode = OTScreenShareVideoViewFit;
+                self.screenSharer.subscriberVideoContentMode = OTVideoViewFit;
             }
         }
     }]];
@@ -104,6 +132,10 @@
 
 -(IBAction)prepareForUnwind:(UIStoryboardSegue *)segue {
     [self.screenSharer updateView:[UIApplication sharedApplication].keyWindow.rootViewController.view];
+}
+
+- (OTAcceleratorSession *)sessionOfOTScreenSharer:(OTScreenSharer *)screenSharer {
+    return [(AppDelegate*)[[UIApplication sharedApplication] delegate] getSharedAcceleratorSession];
 }
 
 @end