+++ ios/ScreenShareAccPackKit/OTScreenShareKit/ScreenShare/OTScreenShareKit.h
@@ -1,7 +1,0 @@
-//
-//  ScreenShareKit.h
-//
-//  Copyright © 2016 Tokbox, Inc. All rights reserved.
-//
-
-#import <OTScreenShareKit/OTScreenSharer.h>