+++ js/build.sh
@@ -1,14 +1,22 @@
 #!/bin/bash
 set -e
+task="$1"
+force="$2"
 
-if [[ -d opentok.js-screen-sharing ]]
-then
+if [ "$task" == "-d" ]; then
+	if [[ -d opentok.js-screen-sharing ]]
+	then
+		cd opentok.js-screen-sharing
+		gulp dist
+		gulp zip
+	else
+		echo "Please run this script from 'JS'."
+		exit 1
+	fi
+fi
+
+if [ "$task" == "-t" ]; then
 	cd opentok.js-screen-sharing
-	gulp dist
-	gulp zip
-  cd dist
-
-else
-	echo "Please run this script from 'JS'."
-	exit 1
+	grunt karma:prod "$force"
+	exit 0
 fi