+++ ios/ScreenShareAccPackKit/CHANGELOG.md
@@ -4,6 +4,51 @@ The changelog for `screensharing-annotation-acc-pack` iOS.
 
 --------------------------------------
 
+### MIGRATION
+
+We will rename and migrate `screen-share-kit` iOS to `OTAcceleratorCore`[here](https://github.com/opentok/accelerator-core-ios)
+
+2.0.5
+-----
+
+### WE ARE MIGRATING THIS PODS TO ANOTHER FOR *GOOD*
+
+2.0.4
+-----
+
+### Enhancements
+
+- Add `isPublishOnly` to allow publishing only.
+
+2.0.3
+-----
+
+### Breaking changes
+
+- Adapt the new `OTAcceleratorPackUtil`.
+
+2.0.2
+-----
+
+### Fixes
+
+- Correct a wrong signaling: `OTPublisherCreated` -> `OTSubscriberReady`.
+
+2.0.1
+-----
+
+### Fixes
+
+- Put back the `isScreenSharing` boolean value.
+
+2.0.0
+-----
+
+### Enhancements
+
+- Remove shared instance so developers can control it.
+- Add OTMultiPartyScreenSharer for mutliparty screen sharing logic.
+
 1.1.6
 -----